requirements.txt
chromadb
sentence-transformers
PyPDF2
python-docx
ollama